+++ python/lsst/ts/wep/task/CalcZernikesTask.py
@@ -110,6 +110,13 @@ class CalcZernikesTaskConfig(
     instPixelSize = pexConfig.Field(
         doc="Instrument Pixel Size in m", dtype=float, default=10.0e-6
     )
+    transposeImages = pexConfig.Field(
+        doc="Specify whether to transpose the intra- and extra-focal images. \
+        (The default is to do the transpose).",
+        dtype=bool,
+        default=True,
+        optional=True,
+    )
 
 
 class CalcZernikesTask(pipeBase.PipelineTask):
@@ -128,7 +135,8 @@ class CalcZernikesTask(pipeBase.PipelineTask):
         # for the detector.
         self.combineZernikes = self.config.combineZernikes
         self.makeSubtask("combineZernikes")
-
+        # Specify whether to transpose images
+        self.transposeImages = self.config.transposeImages
         # Specify optical model
         self.opticalModel = self.config.opticalModel
         # Set up instrument configuration dict
@@ -232,16 +240,23 @@ class CalcZernikesTask(pipeBase.PipelineTask):
             blendOffsetsExtra = self.calcBlendOffsets(donutExtra, eulerZExtra)
             blendOffsetsIntra = self.calcBlendOffsets(donutIntra, eulerZIntra)
 
+            if self.transposeImages:
+                imageExtra = rotate(donutExtra.stamp_im.getImage().getArray(), eulerZExtra).T
+                imageIntra = rotate(donutIntra.stamp_im.getImage().getArray(), eulerZIntra).T
+            else:
+                imageExtra = rotate(donutExtra.stamp_im.getImage().getArray(), eulerZExtra)
+                imageIntra = rotate(donutIntra.stamp_im.getImage().getArray(), eulerZIntra)
+
             wfEsti.setImg(
                 fieldXYExtra,
                 DefocalType.Extra,
-                image=rotate(donutExtra.stamp_im.getImage().getArray(), eulerZExtra).T,
+                image=imageExtra,
                 blendOffsets=blendOffsetsExtra.tolist(),
             )
             wfEsti.setImg(
                 fieldXYIntra,
                 DefocalType.Intra,
-                image=rotate(donutIntra.stamp_im.getImage().getArray(), eulerZIntra).T,
+                image=imageIntra,
                 blendOffsets=blendOffsetsIntra.tolist(),
             )
             wfEsti.reset()